# Producer: Replace Stream with Caching & Parsers: Return Multiple values & TCP Socket Reconnect