--- a/application/apps/indexer/parsers/src/someip.rs
+++ b/application/apps/indexer/parsers/src/someip.rs
@@ -1,11 +1,10 @@
-use crate::{Error, LogMessage, ParseYield, Parser};
+use crate::{parse_all, Error, LogMessage, ParseYield, Parser};
 use std::{
     borrow::Cow,
     cmp::Ordering,
     collections::HashMap,
     fmt::{self, Display},
     io::Write,
-    iter,
     path::PathBuf,
     sync::Mutex,
 };
@@ -26,6 +25,9 @@ use serde::Serialize;
 const COLUMN_SEP: &str = "\u{0004}"; // EOT
 /// Marker for a newline in the output string.
 const LINE_SEP: &str = "\u{0006}"; // ACK
+
+/// The most likely minimal bytes count needed to parse a SomeIP message.
+const MIN_MSG_LEN: usize = 16;
 
 /// Wrapper for a fibex-model (new-type pattern).
 pub struct FibexMetadata {
@@ -271,6 +273,15 @@ impl SomeipParser {
             }
         }
     }
+
+    fn parse_item(
+        &mut self,
+        input: &[u8],
+        timestamp: Option<u64>,
+    ) -> Result<(usize, Option<ParseYield<SomeipLogMessage>>), Error> {
+        SomeipParser::parse_message(self.fibex_metadata.as_ref(), input, timestamp)
+            .map(|(rest, message)| (rest, Some(ParseYield::from(message))))
+    }
 }
 
 unsafe impl Send for SomeipParser {}
@@ -282,10 +293,9 @@ impl Parser<SomeipLogMessage> for SomeipParser {
         input: &[u8],
         timestamp: Option<u64>,
     ) -> Result<impl Iterator<Item = (usize, Option<ParseYield<SomeipLogMessage>>)>, Error> {
-        let item = SomeipParser::parse_message(self.fibex_metadata.as_ref(), input, timestamp)
-            .map(|(rest, message)| (rest, Some(ParseYield::from(message))))?;
-
-        Ok(iter::once(item))
+        parse_all(input, timestamp, MIN_MSG_LEN, |input, timestamp| {
+            self.parse_item(input, timestamp)
+        })
     }
 }
 
@@ -624,7 +634,10 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -645,7 +658,10 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -666,7 +682,10 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -696,7 +715,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -718,7 +740,10 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -749,7 +774,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -774,7 +802,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -799,7 +830,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -824,7 +858,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -849,7 +886,10 @@ mod test {
         let mut parser = SomeipParser {
             fibex_metadata: Some(fibex_metadata),
         };
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -873,7 +913,10 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {
@@ -919,7 +962,11 @@ mod test {
         ];
 
         let mut parser = SomeipParser::new();
-        let (consumed, message) = parser.parse(input, None).unwrap().next().unwrap();
+
+        let mut items: Vec<_> = parser.parse(input, None).unwrap().collect();
+        assert_eq!(items.len(), 1);
+
+        let (consumed, message) = items.pop().unwrap();
         assert_eq!(consumed, input.len());
 
         if let ParseYield::Message(item) = message.unwrap() {