--- a/application/apps/indexer/sources/src/producer/tests/mod.rs
+++ b/application/apps/indexer/sources/src/producer/tests/mod.rs
@@ -1,6 +1,7 @@
 mod mock_byte_source;
 mod mock_parser;
 
+mod cancel_safety;
 mod multi_parse;
 mod single_parse;