--- a/cli/chipmunk-cli/src/main.rs
+++ b/cli/chipmunk-cli/src/main.rs
@@ -1,3 +1,5 @@
+#![warn(unused_extern_crates)]
+
 use tokio::signal;
 use tokio_util::sync::CancellationToken;